# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Command to run the server.",
        "title": "TMS server cmd",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8000",
    "paths": {
        "/batch": {
            "get": {
                "description": "Retrieves a list of batches with optional pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch module"
                ],
                "summary": "Get list of batches",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new batch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch module"
                ],
                "summary": "Create a new batch",
                "parameters": [
                    {
                        "description": "Batch creation details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.CreateBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.Response"
                        }
                    }
                }
            }
        },
...
        "/batch/{id}": {
            "get": {
                "description": "Retrieves details of a batch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch module"
                ],
                "summary": "Get batch details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.BatchResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing batch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch module"
                ],
                "summary": "Update batch details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Batch update details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.UpdateBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an existing batch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch module"
                ],
                "summary": "Delete batch",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.Response"
                        }
                    }
                }
            }
        },
        "/batch/{id}/close": {
            "put": {
                "description": "Closes a batch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch module"
                ],
                "summary": "Close a batch",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.Response"
                        }
                    }
                }
            }
        },
        "/batch/{id}/register": {
            "post": {
                "description": "Registers the user to a batch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch module"
                ],
                "summary": "Register to a batch",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.Response"
                        }
                    }
                }
            }
        },
        "/batch/{id}/registers": {
            "get": {
                "description": "Retrieves a list of students registered to a batch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch module"
                ],
                "summary": "Get list of students registered to a batch",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.BatchInfo": {
            "type": "object",
            "properties": {
                "closed": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.UserInfo"
                },
                "defenceAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "maxTeacherPref": {
                    "type": "integer"
                },
                "maxTeamMember": {
                    "type": "integer"
                },
                "minCGPARequired": {
                    "type": "number"
                },
                "minCHRequired": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "preDefenceAt": {
                    "type": "string"
                },
                "quota": {
                    "type": "string"
                },
                "teamRegDeadline": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.BatchResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.BatchInfo"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.CreateBatchRequest": {
            "type": "object",
            "required": [
                "defenceAt",
                "maxTeacherPref",
                "maxTeamMember",
                "minCHRequired",
                "name",
                "preDefenceAt",
                "teamRegDeadline"
            ],
            "properties": {
                "defenceAt": {
                    "type": "string"
                },
                "maxTeacherPref": {
                    "type": "integer"
                },
                "maxTeamMember": {
                    "type": "integer"
                },
                "minCGPARequired": {
                    "type": "number"
                },
                "minCHRequired": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "preDefenceAt": {
                    "type": "string"
                },
                "quota": {
                    "type": "string"
                },
                "teamRegDeadline": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.StudentInfo": {
            "type": "object",
            "properties": {
                "alternateMobile": {
                    "type": "string"
                },
                "batch": {
                    "type": "integer"
                },
                "cgpa": {
                    "type": "number"
                },
                "completedCredits": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "section": {
                    "type": "string"
                },
                "universityId": {
                    "type": "string"
                }
            }
        },
        "github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.StuffInfo": {
            "type": "object",
            "properties": {
                "alternateMobile": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                }
            }
        },
        "github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.TeacherInfo": {
            "type": "object",
            "properties": {
                "alternateMobile": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "designation": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                }
            }
        },
        "github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.UpdateBatchRequest": {
            "type": "object",
            "properties": {
                "defenceAt": {
                    "type": "string"
                },
                "maxTeacherPref": {
                    "type": "integer"
                },
                "maxTeamMember": {
                    "type": "integer"
                },
                "minCGPARequired": {
                    "type": "number"
                },
                "minCHRequired": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "preDefenceAt": {
                    "type": "string"
                },
                "quota": {
                    "type": "string"
                },
                "teamRegDeadline": {
                    "type": "string"
                }
            }
        },
        "github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.UserInfo": {
            "type": "object",
            "properties": {
                "changePass": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "isAdmin": {
                    "type": "boolean"
                },
                "isVerified": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "student": {
                    "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.StudentInfo"
                },
                "stuff": {
                    "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.StuffInfo"
                },
                "teacher": {
                    "$ref": "#/definitions/github_com_Imtiaz246_Thesis-Management-System_internal_apis_v1.TeacherInfo"
                },
                "universityId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}